requirements.txt
fastapi
uvicorn
openai>=1.0
psycopg2-binary
loguru
prometheus_client
python-dotenv
pytest
psycopg2
loguru
prometheus_client
faiss-cpu
sentence-transformers